requirements.txt
asgiref==3.7.2
Django==4.2.7
django-localflavor==4.0
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
python-stdnum==1.19
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
